Move mock files.
