+++ src/uniprotkb/adapters/__tests__/ExternalLinksConverter.spec.tsx
@@ -2,7 +2,7 @@ import convertExternalLinks from '../externalLinksConverter';
 import { convertXrefProperties } from '../uniProtkbConverter';
 
 import modelData from '../../__mocks__/uniProtKBEntryModelData';
-import databaseInfoMaps from '../__mocks__/databaseInfoMaps';
+import databaseInfoMaps from '../../utils/__tests__/__mocks__/databaseInfoMaps';
 
 describe('External links data converter', () => {
   test('should convert the data', () => {
