+++ src/uniprotkb/adapters/__tests__/FamilyAndDomainsConverter.spec.tsx
@@ -2,7 +2,7 @@ import convertFamilyAndDomains from '../familyAndDomainsConverter';
 import { convertXrefProperties } from '../uniProtkbConverter';
 
 import modelData from '../../__mocks__/uniProtKBEntryModelData';
-import databaseInfoMaps from '../__mocks__/databaseInfoMaps';
+import databaseInfoMaps from '../../utils/__tests__/__mocks__/databaseInfoMaps';
 
 describe('Family and Domains data converter', () => {
   test('should convert the data', () => {
