+++ src/uniprotkb/adapters/__tests__/functionConverter.spec.ts
@@ -2,7 +2,7 @@ import convertFunction from '../functionConverter';
 import { convertXrefProperties } from '../uniProtkbConverter';
 
 import modelData from '../../__mocks__/uniProtKBEntryModelData';
-import databaseInfoMaps from '../__mocks__/databaseInfoMaps';
+import databaseInfoMaps from '../../utils/__tests__/__mocks__/databaseInfoMaps';
 
 const data = convertFunction(
   modelData,
