+++ src/uniprotkb/adapters/__tests__/interactionConverter.spec.ts
@@ -2,7 +2,7 @@ import convertInteraction from '../interactionConverter';
 import { convertXrefProperties } from '../uniProtkbConverter';
 
 import modelData from '../../__mocks__/uniProtKBEntryModelData';
-import databaseInfoMaps from '../__mocks__/databaseInfoMaps';
+import databaseInfoMaps from '../../utils/__tests__/__mocks__/databaseInfoMaps';
 
 describe('Interaction data converter', () => {
   test('should convert the data', () => {
