+++ src/uniprotkb/adapters/__tests__/namesAndTaxonomyConverter.spec.ts
@@ -2,7 +2,7 @@ import { convertNamesAndTaxonomy } from '../namesAndTaxonomyConverter';
 import { convertXrefProperties } from '../uniProtkbConverter';
 
 import modelData from '../../__mocks__/uniProtKBEntryModelData';
-import databaseInfoMaps from '../__mocks__/databaseInfoMaps';
+import databaseInfoMaps from '../../utils/__tests__/__mocks__/databaseInfoMaps';
 
 describe('Names and taxonomy data converter', () => {
   test('should convert the data', () => {
