+++ src/uniprotkb/adapters/__tests__/pathologyAndBiotechConverter.spec.ts
@@ -2,7 +2,7 @@ import diseaseAndDrugs from '../diseaseAndDrugs';
 import { convertXrefProperties } from '../uniProtkbConverter';
 
 import modelData from '../../__mocks__/uniProtKBEntryModelData';
-import databaseInfoMaps from '../__mocks__/databaseInfoMaps';
+import databaseInfoMaps from '../../utils/__tests__/__mocks__/databaseInfoMaps';
 
 describe('Disease and Drugs data converter', () => {
   test('should convert the data', () => {
