+++ src/uniprotkb/adapters/__tests__/structureConverter.spec.ts
@@ -2,7 +2,7 @@ import convertStructure from '../structureConverter';
 import { convertXrefProperties } from '../uniProtkbConverter';
 
 import modelData from '../../__mocks__/uniProtKBEntryModelData';
-import databaseInfoMaps from '../__mocks__/databaseInfoMaps';
+import databaseInfoMaps from '../../utils/__tests__/__mocks__/databaseInfoMaps';
 
 describe('Structure data converter', () => {
   test('should convert the data', () => {
