+++ src/uniprotkb/utils/__tests__/__mocks__/databaseInfoMaps.ts
@@ -1,10 +1,11 @@
-import { DatabaseInfoMaps } from '../../utils/database';
 import databaseCategoryToNames from './databaseCategoryToNames';
 import databaseNameToCategory from './databaseNameToCategory';
 import databaseToDatabaseInfo from './databaseToDatabaseInfo';
 import entrySectionToDatabaseCategoryOrder from './entrySectionToDatabaseCategoryOrder';
 import entrySectionToDatabaseNames from './entrySectionToDatabaseNames';
 import implicitDatabaseXRefs from './implicitDatabaseXrefs';
+
+import { DatabaseInfoMaps } from '../../database';
 
 /**
  * This is generated, based on data from '/configure/uniprotkb/allDatabases'
