+++ src/uniprotkb/utils/__tests__/__mocks__/databaseNameToCategory.ts
@@ -1,4 +1,4 @@
-import { DatabaseCategory } from '../../types/databaseRefs';
+import { DatabaseCategory } from '../../../types/databaseRefs';
 
 const databaseNameToCategory: Map<string, DatabaseCategory> = new Map([
   ['EMBL', DatabaseCategory.SEQUENCE],
