+++ src/uniprotkb/utils/__tests__/__mocks__/databaseToDatabaseInfo.ts
@@ -1,4 +1,4 @@
-import { DatabaseToDatabaseInfo } from '../../utils/database';
+import { DatabaseToDatabaseInfo } from '../../database';
 
 const databaseToDatabaseInfo: DatabaseToDatabaseInfo = {
   EMBL: {
