+++ src/uniprotkb/utils/__tests__/xrefUtils.spec.ts
@@ -6,8 +6,7 @@ import {
   getUnconditionalImplicitXrefs,
   getJoinedXrefs,
 } from '../xrefUtils';
-
-import implicitDatabaseXRefs from '../../adapters/__mocks__/implicitDatabaseXrefs';
+import implicitDatabaseXRefs from './__mocks__/implicitDatabaseXrefs';
 
 describe('xrefUtils tests', () => {
   test('should getDRImplicitXrefs', () => {
